All: reduce warnings (150 -> 25)
This reduces warning with `--Wall --Wextra` on gcc13 from 150 to 25.

The rest are signed issues, which I need to understand how to properly
fix, or unimplemented issues.

+++ src/stackwidget.cpp
@@ -22,7 +22,7 @@ struct SingleWidgetVisible : LayoutItem {
             return;
         }
 
-        auto i = 0;
+        auto i = size_t(0);
         for (auto &item : sub_items) {
             auto w = std::dynamic_pointer_cast<Widget>(item);
             if (w == nullptr) {
